beta svでhost, portの検証を実施
utilityにhost, portの検証コードを移設
